handle PN_XNUM and SHN_XINDEX extended counts
Binaries with >=0xff00 sections or >=0xffff program headers store the real
counts in section 0's header (sh_info/sh_size/sh_link) and put escape values
in the ELF header. The parser read e_phnum/e_shnum/e_shstrndx literally, so
it would mis-size the tables or treat the escape as a real index.

Resolve the escapes from section 0 on read, widen shstrndx to u32, and
re-emit the escapes on write (section 0's extended fields round-trip
untouched).

+++ src/ir.rs
@@ -31,7 +31,9 @@ pub struct Ehdr {
     pub ehsize: u16,
     pub phentsize: u16,
     pub shentsize: u16,
-    pub shstrndx: u16,
+    /// Resolved section-name string-table index; widened to hold an
+    /// SHN_XINDEX value taken from section 0's sh_link.
+    pub shstrndx: u32,
     pub os_abi: u8,
     pub abi_version: u8,
     /// First 16 bytes verbatim; preserves padding/OS-specific bytes on re-encode.
